client:improving auth provider
