+++ client/src/providers/AuthProvider/AuthContext.ts
@@ -1,4 +1,4 @@
-"use client"
+'use client';
 
 import { User } from '@/global_types';
 import { createContext } from 'react';
